fixed all compiler warnings, at least one of which was likely to be a real bug (rt_linux.c ClonePacket function), per @schepanovsky's suggestion

+++ common/mlme.c
@@ -560,8 +560,8 @@ VOID MlmeResetRalinkCounters(
 
 		/* for performace enchanement */
 		NdisZeroMemory(&pAd->RalinkCounters,
-						(UINT32)&pAd->RalinkCounters.OneSecEnd -
-						(UINT32)&pAd->RalinkCounters.OneSecStart);
+						(ULONG)&pAd->RalinkCounters.OneSecEnd -
+						(ULONG)&pAd->RalinkCounters.OneSecStart);
 
 	return;
 }
